Refactored: Management of ZIPed repositories into more generic test base.
Preparation to handle all repositories the same way.
