+++ tests/Arbit/VCSWrapper/GitCli/RepositoryBaseTest.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace Arbit\VCSWrapper\GitCli;
+
+use Arbit\VCSWrapper\RepositoryPreparingBaseTest;
+
+class RepositoryBaseTest extends RepositoryPreparingBaseTest
+{
+    protected static function getVcsIdentifier()
+    {
+        return 'git';
+    }
+}
