+++ tests/bootstrap.php
@@ -1,6 +1,7 @@
 <?php
 
 require_once __DIR__ . '/Arbit/VCSWrapper/TestCase.php';
-require_once __DIR__ . '/Arbit/VCSWrapper/GitCli/RepositoryPreparingBaseTest.php';
+require_once __DIR__ . '/Arbit/VCSWrapper/RepositoryPreparingBaseTest.php';
+require_once __DIR__ . '/Arbit/VCSWrapper/GitCli/RepositoryBaseTest.php';
 require_once __DIR__ . '/../src/library/autoload.php';
